Added 'DriverCarSLBlinkRPM' alias for ShiftLight
'DriverCarSLBlinkRPM' is listed in the SDK documentation as an alternate
to the deprecated 'ShiftIndiactorPct' value for the shift lights. The
documentation is not particularly clear on how to use this (hint: it
requires other YAML data), so it doesn't hurt to have the string as an
alias to the custom function.